Quiet windows build warnings.
svn:revision:59762
svn:branch:trunk
svn:account:r_weiss

+++ src/rt/view.c
@@ -927,7 +927,7 @@ ambientOcclusion(struct application *ap, struct partition *pp)
     vect_t uAxis;
     int ao_samp;
     vect_t origin = VINIT_ZERO;
-    float occlusionFactor;
+    double occlusionFactor;
     int hitCount = 0;
 
     stp = pp->pt_inseg->seg_stp;
